+++ IsometricEngine/src/Tree.h
@@ -2,7 +2,6 @@
 #include "Scenery.h"
 class Tree : public Scenery
 {
-public:
-    Tree(int initialPosX,int initialPosY):Scenery(initialPosX,initialPosY,"Tree"){}
+   public:
+	Tree(float initialPosX, float initialPosY) : Scenery(initialPosX, initialPosY, "Tree") {}
 };
-
